+++ packages/@baldeweg/ui/src/components/BFlex/BFlex.spec.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-import BFlex from '../BFlex/BFlex.vue'
-
-describe('BFlex', () => {
-  it('shows BFlex', () => {
-    expect(BFlex).toBeTruthy()
-  })
-})