+++ packages/@baldeweg/ui/src/components/BAlert/BAlert.vue
@@ -4,49 +4,40 @@ import { ref, computed } from 'vue'
 const props = defineProps({
   type: {
     type: String,
-    validator: (value) =>  ['neutral', 'info', 'danger', 'warning', 'success'].includes(value),
-    default: 'neutral',
+    validator: (value) => ['info', 'danger', 'warning', 'success'].includes(value),
+    default: 'info',
   },
   closable: {
     type: Boolean,
     default: false
-  },
-  position: {
-    type: String,
-    validator: (value) => ['block', 'top', 'bottom', 'left', 'right'].includes(value),
-    default: 'block',
   },
 })
 
 const showAlert = ref(false)
 const alertClass = computed(() => {
   return {
-    alert_neutral: props.type === 'neutral',
     alert_info: props.type === 'info',
     alert_danger: props.type === 'danger',
     alert_warning: props.type === 'warning',
     alert_success: props.type === 'success',
-    alert_block: props.position === 'block',
-    alert_top: props.position === 'top' ,
-    alert_bottom: props.position === 'bottom' ,
-    alert_topleft: props.position === 'topleft' ,
-    alert_topright: props.position === 'topright' ,
-    alert_bottomleft: props.position === 'bottomleft',
-    alert_bottomright: props.position === 'bottomright',
   }
 })
 </script>
 
 <template>
   <div class="alert" :class="alertClass" v-if="!showAlert">
-    <BMaterialIcon :size="28" @click="showAlert = true" hover :style="{ float: 'right' }" v-if="closable">close
-    </BMaterialIcon>
-    <slot />
+    <BMaterialIcon :size="20" class="alert_icon">info</BMaterialIcon>
+    <div class="alert-content">
+      <slot />
+    </div>
+    <BMaterialIcon :size="20" @click="showAlert = true" hover class="alert_close" v-if="closable">close</BMaterialIcon>
   </div>
 </template>
 
 <style>
 .alert {
+  display: flex;
+  gap: 20px;
   border-radius: 20px;
   border: 1px solid var(--color-neutral-08);
   background: var(--color-neutral-00);
@@ -55,67 +46,56 @@ const alertClass = computed(() => {
   box-sizing: border-box;
 }
 
-/* Type */
-.alert_neutral {
-  border-color: var(--color-neutral-08);
-}
-.alert_info {
-  border-color: var(--color-accent-blue-10);
-}
-.alert_danger {
-  border-color: var(--color-accent-red-10);
-}
-.alert_warning {
-  border-color: var(--color-accent-yellow-10);
-}
-.alert_success {
-  border-color: var(--color-accent-green-10);
+.alert_icon {
+  user-select: none;
 }
 
-/* Position */
-.alert_block {
+.alert-content {
+  flex: 1;
 }
-.alert_top,
-.alert_bottom,
-.alert_topleft,
-.alert_topright,
-.alert_bottomleft,
-.alert_bottomright {
-   position: fixed;
-   border-radius: 20px;
-   width: 400px;
-   z-index: 2;
+
+/* Type */
+.alert_info {
+  border-color: var(--color-accent-blue-10);
+  color: var(--color-accent-blue-10)
 }
-.alert_top {
-  top: 0;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
+
+.alert_info a,
+.alert_info a:hover {
+  color: var(--color-accent-blue-10);
+  text-decoration-color: var(--color-accent-blue-10);
 }
-.alert_bottom {
-  bottom: 0;
-  left: 0;
-  right: 0;
-  margin: 0 auto;
+
+.alert_danger {
+  border-color: var(--color-accent-red-10);
+  color: var(--color-accent-red-10);
 }
-.alert_topleft {
-  top: 0;
-  left: 0;
-  margin: 0;
+
+.alert_danger a,
+.alert_danger a:hover {
+  color: var(--color-accent-red-10);
+  text-decoration-color: var(--color-accent-red-10);
 }
-.alert_topright {
-  top: 0;
-  right: 0;
-  margin: 0;
+
+.alert_warning {
+  border-color: var(--color-accent-yellow-10);
+  color: var(--color-accent-yellow-10)
 }
-.alert_bottomleft {
-  bottom: 0;
-  left: 0;
-  margin: 0;
+
+.alert_warning a,
+.alert_warning a:hover {
+  color: var(--color-accent-yellow-10);
+  text-decoration-color: var(--color-accent-yellow-10);
 }
-.alert_bottomright {
-  bottom: 0;
-  right: 0;
-  margin: 0;
+
+.alert_success {
+  border-color: var(--color-accent-green-10);
+  color: var(--color-accent-green-10)
+}
+
+.alert_success a,
+.alert_success a:hover {
+  color: var(--color-accent-green-10);
+  text-decoration-color: var(--color-accent-green-10);
 }
 </style>